Add support for SSH key authentication
This commit adds support for SSH key authentication. Users can now specify multiple public keys that can be used for authentication. The keys are specified in the user configuration file using the `ssh_keys` field. Each key is specified with its type and base64 string.

Note: Authentication logic just checks if user is associated with a key
provided. The validity of the key is checked by libssh.